Add support for FreeBSD
It's the same logic as with hidraw under Linux, so sprinkle some ifdef
magic around.

While here, have a defaults and overrides section in the Makefile; fix a
missing $(CC) variable and also default it to 'cc', which usually does
the right thing with gcc/clang depending on the user's preference.